pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linksentry"
version = "0.1.0"
description = "Explainable URL and HTML phishing analyzer"
authors = [{ name = "LinkSentry" }]
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
dependencies = [
    "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0"
]

[project.scripts]
linksentry = "linksentry.cli:main"

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py"]
addopts = "-q"
